# docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/fleets/{fleetId}/trips": {
      "post": {
        "operationId": "TripsController_create",
        "parameters": [
          {
            "name": "fleetId",
            "required": true,
            "in": "path",
            "description": "Fleet selector (uuid) — a selector only; membership is verified server-side.",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTripDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Trip created (defaults to SCHEDULED).",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "statusCode": {
                      "type": "integer",
                      "example": 201
                    },
                    "data": {
                      "$ref": "#/components/schemas/TripDto"
                    }
                  },
                  "required": [
                    "statusCode",
                    "data"
                  ]
                }
              }
            }
          },
          "401": {
            "description": "Missing/invalid/expired token, revoked session, stale authVersion, or deactivated account."
          },
          "403": {
            "description": "Authenticated but lacking the required permission (or no ACTIVE membership)."
          },
          "404": {
            "description": "Bus not found in this fleet (cross-fleet bus ids are also 404)."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Schedule a trip with one of the fleet buses.",
        "tags": [
          "trips"
        ]
      },
      "get": {
        "operationId": "TripsController_list",
        "parameters": [
          {
            "name": "fleetId",
            "required": true,
            "in": "path",
            "description": "Fleet selector (uuid) — a selector only; membership is verified server-side.",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "cursor",
            "required": false,
            "in": "query",
            "description": "Opaque nextCursor from the previous page (base64url of the last item id).",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Page size (default 20, max 100).",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 20
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Cursor page of trips (items + nextCursor).",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "statusCode": {
                      "type": "integer",
                      "example": 200
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/TripDto"
                      }
                    }
                  },
                  "required": [
                    "statusCode",
                    "data"
                  ]
                }
              }
            }
          },
          "401": {
            "description": "Missing/invalid/expired token, revoked session, stale authVersion, or deactivated account."
          },
          "403": {
            "description": "Authenticated but lacking the required permission (or no ACTIVE membership)."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List the fleet trips (cursor pagination).",
        "tags": [
          "trips"
        ]
      }
    },
    "/fleets/{fleetId}/trips/{id}": {
      "get": {
        "operationId": "TripsController_findOne",
        "parameters": [
          {
            "name": "fleetId",
            "required": true,
            "in": "path",
            "description": "Fleet selector (uuid) — a selector only; membership is verified server-side.",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Trip id (uuid).",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The trip.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "statusCode": {
                      "type": "integer",
                      "example": 200
                    },
                    "data": {
                      "$ref": "#/components/schemas/TripDto"
                    }
                  },
                  "required": [
                    "statusCode",
                    "data"
                  ]
                }
              }
            }
          },
          "401": {
            "description": "Missing/invalid/expired token, revoked session, stale authVersion, or deactivated account."
          },
          "403": {
            "description": "Authenticated but lacking the required permission (or no ACTIVE membership)."
          },
          "404": {
            "description": "Trip not found in this fleet (cross-fleet ids are also 404)."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Fetch one trip.",
        "tags": [
          "trips"
        ]
      },
      "patch": {
        "operationId": "TripsController_update",
        "parameters": [
          {
            "name": "fleetId",
            "required": true,
            "in": "path",
            "description": "Fleet selector (uuid) — a selector only; membership is verified server-side.",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Trip id (uuid).",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateTripDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated trip.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "statusCode": {
                      "type": "integer",
                      "example": 200
                    },
                    "data": {
                      "$ref": "#/components/schemas/TripDto"
                    }
                  },
                  "required": [
                    "statusCode",
                    "data"
                  ]
                }
              }
            }
          },
          "401": {
            "description": "Missing/invalid/expired token, revoked session, stale authVersion, or deactivated account."
          },
          "403": {
            "description": "Authenticated but lacking the required permission (or no ACTIVE membership)."
          },
          "404": {
            "description": "Trip not found in this fleet (cross-fleet ids are also 404)."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update a trip (route/timing/status).",
        "tags": [
          "trips"
        ]
      },
      "delete": {
        "operationId": "TripsController_remove",
        "parameters": [
          {
            "name": "fleetId",
            "required": true,
            "in": "path",
            "description": "Fleet selector (uuid) — a selector only; membership is verified server-side.",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Trip id (uuid).",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Trip deleted; data is null.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "statusCode": {
                      "type": "integer",
                      "example": 200
                    },
                    "data": {
                      "nullable": true,
                      "example": null
                    }
                  },
                  "required": [
                    "statusCode",
                    "data"
                  ]
                }
              }
            }
          },
          "401": {
            "description": "Missing/invalid/expired token, revoked session, stale authVersion, or deactivated account."
          },
          "403": {
            "description": "Authenticated but lacking the required permission (or no ACTIVE membership)."
          },
          "404": {
            "description": "Trip not found in this fleet (cross-fleet ids are also 404)."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Delete a trip.",
        "tags": [
          "trips"
        ]
      }
    },
...
  },
  "info": {
    "title": "Bus Fleet API",
    "description": "Multi-tenant group bus transportation platform REST API.\n\n- **Authentication is owned by this API**: HS256 JWTs carry a custom `app_role` claim; refresh tokens rotate on every use and are stored hashed at rest.\n- **Every response is wrapped in a `{ statusCode, data }` envelope**; the `data` schema is shown inside each documented response.\n- **Fleet isolation is enforced by PostgreSQL RLS** as the final security boundary; cross-tenant lookups are indistinguishable from missing rows and return 404.\n- `fleetId` path parameters are **selectors, never proof of authorization** — access is verified against the caller's ACTIVE `FleetMember` row on every request.\n- Roles and permissions are **database-driven** (no hardcoded enums); only `super_admin` is a predefined system role. Platform routes (roles, permissions, users, fleets admin, audit-logs) are privileged: super admin only and audited.",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "TripDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "fleetId": {
            "type": "string",
            "format": "uuid"
          },
          "busId": {
            "type": "string",
            "format": "uuid"
          },
          "origin": {
            "type": "string",
            "example": "Cairo"
          },
          "destination": {
            "type": "string",
            "example": "Alexandria"
          },
          "departAt": {
            "format": "date-time",
            "type": "string",
            "example": "2026-09-10T08:00:00.000Z"
          },
          "status": {
            "type": "string",
            "enum": [
              "SCHEDULED",
              "DEPARTED",
              "COMPLETED",
              "CANCELLED"
            ],
            "example": "SCHEDULED"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "fleetId",
          "busId",
          "origin",
          "destination",
          "departAt",
          "status",
          "createdAt",
          "updatedAt"
        ]
      },
      "CreateTripDto": {
        "type": "object",
        "properties": {
          "busId": {
            "type": "string",
            "format": "uuid",
            "description": "Bus id — must belong to the same fleet (cross-fleet bus ids resolve to 404)."
          },
          "origin": {
            "type": "string",
            "example": "Cairo",
            "minLength": 1,
            "maxLength": 255
          },
          "destination": {
            "type": "string",
            "example": "Alexandria",
            "minLength": 1,
            "maxLength": 255
          },
          "departAt": {
            "type": "string",
            "example": "2026-09-10T08:00:00.000Z",
            "format": "date-time"
          },
          "status": {
            "type": "string",
            "enum": [
              "SCHEDULED",
              "DEPARTED",
              "COMPLETED",
              "CANCELLED"
            ],
            "default": "SCHEDULED"
          }
        },
        "required": [
          "busId",
          "origin",
          "destination",
          "departAt"
        ]
      },
      "UpdateTripDto": {
        "type": "object",
        "properties": {
          "origin": {
            "type": "string",
            "example": "Cairo",
            "minLength": 1,
            "maxLength": 255
          },
          "destination": {
            "type": "string",
            "example": "Alexandria",
            "minLength": 1,
            "maxLength": 255
          },
          "departAt": {
            "type": "string",
            "example": "2026-09-10T08:00:00.000Z",
            "format": "date-time"
          },
          "status": {
            "type": "string",
            "enum": [
              "SCHEDULED",
              "DEPARTED",
              "COMPLETED",
              "CANCELLED"
            ]
          }
        }
      },
...
    }
  }
}